shared: Avoid signed extension on multiplication
Depending on factor values, we could end up having a huge number, due to
expansion of signal bit on unsigned conversion.

+++ src/shared/sol-vector.c
@@ -106,7 +106,7 @@ sol_vector_shrink(struct sol_vector *v)
 int
 sol_vector_del(struct sol_vector *v, uint16_t i)
 {
-    int tail_len;
+    size_t tail_len;
 
     if (i >= v->len)
         return -EINVAL;
